\name{commit_job}
\alias{commit_job}
\title{Commit a list of jobs to redis server}
\usage{
  commit_job(job.list, is.processbar = TRUE)
}
\arguments{
  \item{job.list}{a list of jobs to be committed}

  \item{is.processbar}{logical, To determine whether a
  processbar is shown or not}
}
\description{
  Commit a list of jobs to redis server
}

