\name{init_worker}
\alias{init_worker}
\title{init_worker

Ask the server how to initialize}
\usage{
  init_worker(path = NULL, shared_secret = "default")
}
\arguments{
  \item{path}{the path of server, ex:
  "tcp://localhost:12345"}

  \item{shared_secret}{a secrete share with the server.}
}
\description{
  init_worker

  Ask the server how to initialize
}

