\name{pop_job_processing}
\alias{pop_job_processing}
\title{pop_job_processing

Return the job in the hash values in redis according to the parameter \code{hash}}
\usage{
  pop_job_processing(hash)
}
\arguments{
  \item{hash}{the hash value of the job}
}
\value{
  an instance of 'job'
}
\description{
  pop_job_processing

  Return the job in the hash values in redis according to
  the parameter \code{hash}
}

